requirements.txt
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
django-jazzmin==2.4.9
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
importlib-metadata==4.11.3
mysqlclient==2.1.0
Pillow==9.0.1
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
zipp==3.7.0
